// Qanat.Web/src/app/shared/generated/model/reporting-period-cover-crop-self-report-metadata-update-dto.ts
/**
 * Qanat.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ReportingPeriodCoverCropSelfReportMetadataUpdateDto { 
    CoverCropSelfReportStartDate?: string | null;
    CoverCropSelfReportEndDate?: string | null;
    CoverCropSelfReportReadyForAccountHolders?: boolean | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ReportingPeriodCoverCropSelfReportMetadataUpdateDtoForm { 
    CoverCropSelfReportStartDate?: FormControl<string>;
    CoverCropSelfReportEndDate?: FormControl<string>;
    CoverCropSelfReportReadyForAccountHolders?: FormControl<boolean>;
}

export class ReportingPeriodCoverCropSelfReportMetadataUpdateDtoFormControls { 
    public static CoverCropSelfReportStartDate = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CoverCropSelfReportEndDate = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CoverCropSelfReportReadyForAccountHolders = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}